Split list item tag and spacer
